+++ app/src/main/res/layout/activity_main.xml
@@ -13,6 +13,7 @@
         style="@style/DoneButton"
         android:layout_centerHorizontal="true"
         android:layout_marginTop="10dp" />
+    
 
 
     <ProgressBar
